// src/app/enums/ESharedInputType.ts
export enum ESharedInputType {
  DESTINATIONS_FLIGHTS = 'DESTINATIONS_FLIGHTS',
  DESTINATIONS_FLIGHTS_EILAT = 'DESTINATIONS_FLIGHTS_EILAT',
  MULTI_DESTINATION_ORIGINS_FLIGHTS = 'MULTI_DESTINATION_ORIGINS_FLIGHTS',
  HOTELS_DESTINATION = 'HOTELS_DESTINATION',
  DESTINATIONS_FLIGHTS_HOTEL_EILAT = 'DESTINATIONS_FLIGHTS_HOTEL_EILAT',
  DOMESTIC_VACATION_DESTINATION = 'DOMESTIC_VACATION_DESTINATION',
  ORIGINS_FLIGHTS = 'ORIGINS_FLIGHTS',
  ORIGINS_FLIGHTS_EILAT = 'ORIGINS_FLIGHTS_EILAT',
  PASSANGERS_FLIGHTS = 'PASSANGERS_FLIGHTS',
  PASSANGERS_ABOARD_HOTEL = 'PASSANGERS_ABOARD_HOTEL',
  PASSANGERS_DOMESTIC_VACATION = 'PASSANGERS_DOMESTIC_VACATION',
  PASSANGERS_FLIGHTS_EILAT = 'PASSANGERS_FLIGHTS_EILAT',
  PASSANGERS_FLIGHTS_AND_HOTEL_EILAT = 'PASSANGERS_FLIGHTS_AND_HOTEL_EILAT',
  PICKER_DATES = 'PICKER_DATES',
  SPORT_EVENT_TYPE = 'SPORT_EVENT_TYPE',
  SPORT_LEAGUES = 'SPORT_LEAGUES',
  SPORT_TEAMS = 'SPORT_TEAMS',
  ORGANIZED_REGIONS = 'ORGANIZED_REGIONS',
  ORGANIZED_COUNTRIES = 'ORGANIZED_COUNTRIES',
  ORGANIZED_CATEGORIES = 'ORGANIZED_CATEGORIES',
  DATES_PICKER_MONTHS = 'DATES_PICKER_MONTHS',
  DYNAMIC_PACKAGES_DESTINATION = 'DYNAMIC_PACKAGES_DESTINATION',
  DYNAMIC_PACKAGES_DATES = 'DYNAMIC_PACKAGES_DATES',
  PASSANGERS_DYNAMIC_PACKAGES = 'PASSANGERS_DYNAMIC_PACKAGES',
  SKI_DESTINATION = 'SKI_DESTINATION',
  SKI_RESORT = 'SKI_RESORT',
  SKI_DEPARTURE_DATE = 'SKI_DEPARTURE_DATE',
  SKI_PASSENGERS = 'SKI_PASSENGERS',
  VILLAGE_RESORTS_DESTINATION = 'VILLAGE_RESORTS_DESTINATION',
  VILLAGE_RESORTS_DATES = 'VILLAGE_RESORTS_DATES',
  VILLAGE_RESORTS_PASSENGERS = 'VILLAGE_RESORTS_PASSENGERS',
  CAR_PICKUP_COUNTRY = 'CAR_PICKUP_COUNTRY',
  CAR_PICKUP_CITY = 'CAR_PICKUP_CITY',
  CAR_DATES = 'CAR_DATES',
  CAR_DRIVER_AGE = 'CAR_DRIVER_AGE',
}